docs: update Fractal version support
Changes have been tested in a fairly large (~1100 component variants) Fractal project. One potential issue is compatibility with the deprecated syntax of more than one `@`. Like:

@if (@Model.foo) {}

Vash is supposed to be backward-compatible but I ran in to circumstances. I didn't manage to fully track the issue down, but I believe it was introduced in vash@0.12.3.